More cleanup and clippy fix
