+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod block_iter;
 pub mod decode;
 pub mod encode;
 pub mod types;
@@ -5,7 +6,7 @@ pub mod util;
 
 // TODO: use larger seeds for the PRNG
 // TODO: allow specification of starting block_id
-// TODO: write tests with proptest
+// TODO: write more tests with proptest
 // TODO: write benchmarks with criterion
 // TODO: profile and fix low-hanging fruit
 // TODO: write docs
